antispoofing-backend/auth-service/requirements.txt
confluent-kafka
opencv-python-headless
django-environ
boto3
dlib
face-recognition
